Add feature race condition

+++ src/main/java/com/example/cinemaproject/repository/SeatRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SeatRepository extends JpaRepository<SeatEntity, Integer> {
+public interface SeatRepository extends JpaRepository<SeatEntity, Long> {
     List<SeatEntity> findByRoomId(Long roomId);
 }
